requirements.txt
ultralytics>=8.2.0
opencv-python
numpy
pyyaml
streamlit
pandas
scipy
lapx            # for trackers (ByteTrack/BoT-SORT dep)
onnxruntime     # ONNX inference path
filterpy        # tracker deps
